+++ packages/website/gatsby-browser.js
@@ -1,1 +1,2 @@
-export { wrapPageElement } from './src';
+export { wrapRootElement } from './src/wrapRootElement';
+export { wrapPageElement } from './src/wrapPageElement';
