+++ packages/website/src/wrapRootElement.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import { RadixProvider } from '@modulz/radix';
+
+export const wrapRootElement = ({ element }) => {
+  return <RadixProvider>{element}</RadixProvider>;
+};
